move models into their own mods
